+++ test/test_helper.rb
@@ -7,5 +7,6 @@ end
 
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "inttegro"
+require "opentelemetry-sdk"
 
 require "minitest/autorun"